Just-Dice.com Roll Emulation
Currently, the source "emulates" the HmacSha512 function by using the
encrypted server key; Also it emulates true rolls as would be rolled on
the just-dice.com website. Then it compares each element of the arrays
on 49.5% chance and gives a counter increase if the numbers fall within
the same category.

10,000 rolls gives the assumption that by rolling the opposite roll as
suggested by the emulation function, one has a greater chance of winning
in the short term; Infinite time outlook yeilds 49.5% chance of winning.